fix: Implement IUnoKeyboardInputSource on FrameBuffer
